refactor: router 코드 관심사 분리
